project_files/ai_powered_log_anomaly_detector_20260404_130654/ai_powered_log_anomaly_detector/config/config.py
import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
FAISS_INDEX_PATH = os.getenv("FAISS_INDEX_PATH", "./faiss_index")
MAX_LOGS_PER_REQUEST = 1000
MAX_RETRIES = 3
LOG_LEVEL = "INFO"

---