project_files/ai_powered_log_anomaly_detector_20260404_130654/ai_powered_log_anomaly_detector/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
streamlit==1.26.0
pandas==1.5.3
faiss-cpu==1.7.3
openai==0.28.1
python-dotenv==0.21.1
pydantic==2.5.2
uvicorn==0.21.1
loguru==0.6.0
typing-extensions==4.5.0